refactor: merged forms from EditBook and AddBook into AddEditBookForm
